requirements.txt
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.30.0
schedule==1.2.0
pytz==2023.3